load models directly from github

+++ R/02-downAndUploadModel.R
@@ -96,17 +96,9 @@ uploadModelUI <- function(id, label) {
   
   tagList(
     tags$h5(label),
-    selectInput(
-      ns("remoteModel"),
-      label = "Select remote model",
-      choices = dir(file.path(settings$pathToSavedModels)) %>%
-        sub(pattern = '\\.zip$', replacement = ''),
-      selected = NULL
-    ),
-    actionButton(ns("loadRemoteModel"), "Load Remote Model"),
-    tags$br(), tags$br(),
-    #helpText("Remote models are only available on on https://isomemoapp.com")
-    fileInput(ns("uploadModel"), label = NULL)
+    fileInput(ns("uploadModel"), label = "Load local model"),
+    remoteModelsUI(ns("remoteModels")),
+    tags$br(), tags$br(), tags$br()
   )
 }
 
@@ -135,8 +127,15 @@ uploadModel <- function(input, output, session, savedModels, uploadedNotes, fit,
     pathToModel(input$uploadModel$datapath)
   })
   
-  observeEvent(input$loadRemoteModel, {
-    pathToModel(file.path(settings$pathToSavedModels, paste0(input$remoteModel, ".zip")))
+  pathToRemote <- remoteModelsServer("remoteModels",
+                                     githubRepo = "osteo-bior",
+                                     rPackageName = "OsteoBioR",
+                                     rPackageVersion = "OsteoBioR" %>%
+                                       packageVersion() %>%
+                                       as.character())
+  
+  observeEvent(pathToRemote(), {
+    pathToModel(pathToRemote())
   })
   
   observeEvent(pathToModel(), {
@@ -154,9 +153,9 @@ uploadModel <- function(input, output, session, savedModels, uploadedNotes, fit,
       return()
     }
     
-    file.remove("model.rds")
-    file.remove("README.txt")
-    file.remove("help.html")
+    if (file.exists("model.rds")) file.remove("model.rds")
+    if (file.exists("README.txt")) file.remove("README.txt")
+    if (file.exists("help.html")) file.remove("help.html")
     
     if (!exists("modelImport")) {
       shinyjs::alert("File format not valid. Model object not found.")
